requirements.txt
Django==5.1.5
djangorestframework
django-ckeditor
django-redis
googletrans==4.0.0-rc1
pytest
pytest-django
redis